deploy.sh: adapt to new mlab-deployment style
The -n option is gone, because the code was moved into
the prepare.sh script by Stephen.

I needed to add a workaround for sudo, which refuses to
run in the new slivers unless an askpass program is
provided.

Now there is a unique tarball (mlab_neubot.tar) that
contains both the init/ scripts and the code.

I removed the possibility of fetching automatically the
list of servers. One can always invoke ls.py and pass
the server names to this script.

Finally, I adapted the code to the new style: now, it
just unpacks mlab_neubot.tar and delegates the rest
of the work to init/initialize.sh.